discordbot/requirements.txt
discord.py==2.3.1
python-dotenv==0.20.0
sqlalchemy
aiosqlite
